CustomApplicationUtils split up
